Refactored a lot of code. Moved code over to FitsFile.py class
