+++ fitsfile.py
@@ -7,6 +7,8 @@ import matplotlib.pyplot as plt
 from matplotlib.colors import LogNorm
 import cv2
 import time
+import scipy
+from spectrum import Spectrum
 
 class FitsFile:
     def __init__(self, fits_file):
@@ -24,7 +26,7 @@ class FitsFile:
     def get_dimensions(self):
         return (self.rows, self.cols)
 
-    def plot_spectra(self, spectra, show=False):
+    def plot_spectra(self, show=False):
         """
         Plots the spectra on top of the image.
         """
@@ -32,13 +34,12 @@ class FitsFile:
         # importing inside function to fix circular dependency issue
         import cleanup 
 
-        image = self.image_data
-        thresholded_im = cleanup.threshold_image(image)
+        thresholded_im = cleanup.threshold_image(self.image_data)
         plt.imshow(thresholded_im, origin="lower", cmap="gray")
-        image_rows = image.shape[0]
-        image_cols = image.shape[1]
+        image_rows = self.image_data.shape[0]
+        image_cols = self.image_data.shape[1]
         degree = 3
-        for spectrum in spectra:
+        for spectrum in self.spectra:
             spectrum.plot()
             spectrum.fit_polynomial(np.arange(0, image_cols), degree)
             spectrum.plot_fit()
@@ -50,8 +51,110 @@ class FitsFile:
             plt.show()
 
 
+    def __get_intensity_array(self, xpixel=1000):
+        """
+        At a particular xpixle, (default is 1000), this function returns
+        an array of intensity for each pixel in y.
+        """
+        intensity = []
+        for row_num, row in enumerate(self.image_data):
+            intensity.append(row[xpixel])
+
+        return np.array(intensity)
+
+    
+    def __find_peaks(self, intensity_array):
+        """
+        Find peaks in the intensity array. The peaks correspond to each order of the spectrograph.
+        """
+        # ignores peaks with intensities less than 100
+        peaks = scipy.signal.find_peaks(intensity_array, height=100) 
+        
+        return peaks[0]
+
+
+    def get_spectra(self):
+        
+        # This is the pixel we use to determine how many spectra there are
+        prime_pixel = 1000 
+        xpixels = np.arange(self.image_data.shape[1])
+        print("xpixels:", xpixels)
+
+        list_of_IA = [] # list of intensity arrays
+        list_of_peaks = []  
+
+        # Finds the number of peaks based on the reference prime x pixel
+        num_peaks_prime = self.__find_peaks(self.__get_intensity_array(xpixel=prime_pixel))
+        used_xpixels = []
+
+        # Obtain the y v. intensity for each x pixel so 
+        # we can find peaks for various values in the domain.
+        for xpixel in xpixels:
+            ia = self.__get_intensity_array(xpixel=xpixel)
+            list_of_IA.append(ia)
+            peaks = self.__find_peaks(ia)
+
+            # This helps ensure that the same number of spectra are detected in each
+            # xpixel as the reference xpixel, which is tentatively set to 1000
+            if len(peaks) == len(num_peaks_prime):
+                list_of_peaks.append(np.array(peaks))
+                used_xpixels.append(xpixel)
+
+        # numpifying list_of_peaks
+        # This array contains peaks for each x value 
+        # in x pixels.
+        list_of_peaks = np.array(list_of_peaks)
+
+        # Each order is identified by the index of peak. 
+        # Have to ensure that list_of_peaks is rectangular because
+        # we have to ensure that we are detecting the same number
+        # of spectra for each x value. Otherwise, something will be 
+        # discontinuous.
+        assert(is_rectangular(list_of_peaks))
+
+        spectra = []
+        xvals = [] # An array that contains x values of each pixel.
+
+        for x in used_xpixels:
+            xvals.append(np.repeat(x, len(list_of_peaks[0])))
+        
+        # numpifying array
+        xvals = np.array(xvals)
+
+        # Correctness check
+        assert(np.array(xvals).shape == np.array(list_of_peaks).shape)
+
+        num_cols = len(list_of_peaks[0])
+
+        # list_of_peaks contains the peaks for a singular x value, and 
+        # xvalues contain the same x values for the length of each 
+        # array in list_of_peaks. So this simply takes each column
+        # from each array, which gives us the x and y coordinates
+        # for each spectrum.
+        for i in range(num_cols):
+            xvalues = xvals[:, i]
+            yvalues = list_of_peaks[:, i]
+            spectra.append(Spectrum(xvalues, yvalues))
+
+        self.spectra = spectra
+
+
+
     def get_file_name(self):
         """
         Returns the name of the file independent of the path.
         """
         return self.fits_file[self.fits_file.rfind("/")+1:]
+
+
+
+def is_rectangular(l):
+    """
+    This is a correctness check function that is used in get_spectra() routine below.
+    It ensures that the given array l is rectangular.
+    """
+    for i in l:
+        if len(i) != len(l[0]):
+            return False
+
+    return True
